Rename CM Auto Poster to Content Poster and did major refactoring and code documentation
